Refactor to multi-group
