+++ PROJECT
@@ -5,6 +5,7 @@
 domain: cluster.x-k8s.io
 layout:
 - go.kubebuilder.io/v4
+multigroup: true
 projectName: cluster-api-provider-gardener
 repo: github.com/gardener/cluster-api-provider-gardener
 resources:
