+++ api/controlplane/v1alpha1/gardenershootcontrolplane_types.go
