+++ api/types.go
@@ -1,4 +1,4 @@
-package v1alpha1
+package api
 
 import (
 	gardenercorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
@@ -8,6 +8,9 @@ import (
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 	clusterv1beta1 "sigs.k8s.io/cluster-api/api/v1beta1"
+
+	controlplanev1alpha1 "github.com/gardener/cluster-api-provider-gardener/api/controlplane/v1alpha1"
+	infrastructurev1alpha1 "github.com/gardener/cluster-api-provider-gardener/api/infrastructure/v1alpha1"
 )
 
 var (
@@ -19,7 +22,8 @@ func init() {
 	utilruntime.Must(clientgoscheme.AddToScheme(Scheme))
 
 	utilruntime.Must(clusterv1beta1.AddToScheme(Scheme))
-	utilruntime.Must(AddToScheme(Scheme))
+	utilruntime.Must(controlplanev1alpha1.AddToScheme(Scheme))
+	utilruntime.Must(infrastructurev1alpha1.AddToScheme(Scheme))
 
 	utilruntime.Must(gardenercorev1beta1.AddToScheme(Scheme))
 	utilruntime.Must(kubernetes.AddGardenSchemeToScheme(Scheme))
